Add fdroid flavor
