+++ app/src/fdroid/java/com/superproductivity/superproductivity/JavaScriptInterface.java
@@ -1,0 +1,12 @@
+package com.superproductivity.superproductivity;
+
+import android.webkit.WebView;
+
+public class JavaScriptInterface extends CommonJavaScriptInterface {
+    /**
+     * Instantiate the interface and set the context
+     */
+    JavaScriptInterface(FullscreenActivity c, WebView wv) {
+        super(c, wv);
+    }
+}
